+++ crates/miden-objects/src/asset/vault/partial.rs
@@ -1,16 +1,19 @@
+use alloc::string::ToString;
+
 use miden_crypto::merkle::{InnerNodeInfo, MerkleError, PartialSmt, SmtLeaf, SmtProof};
-use vm_core::utils::{Deserializable, Serializable};
 
 use super::AssetVault;
 use crate::Word;
 use crate::asset::Asset;
+use crate::errors::PartialAssetVaultError;
+use crate::utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
 
 /// A partial representation of an [`AssetVault`], containing only proofs for a subset of assets.
 ///
 /// Partial vault is used to provide verifiable access to specific assets in a vault
 /// without the need to provide the full vault data. It contains all required data for loading
 /// vault data into the transaction kernel for transaction execution.
-#[derive(Clone, Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, Default)]
 pub struct PartialVault {
     /// An SMT with a partial view into an account's full [`AssetVault`].
     partial_smt: PartialSmt,
@@ -20,10 +23,18 @@ impl PartialVault {
     // CONSTRUCTORS
     // --------------------------------------------------------------------------------------------
 
-    // TODO: Do we need this constructor? It does not validate that the SMT contains valid Assets.
-    /// Returns a new instance of partial vault with the specified root and vault proofs.
-    pub fn new(partial_smt: PartialSmt) -> Self {
-        PartialVault { partial_smt }
+    /// Returns a new instance of a partial vault from the provided partial SMT.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if:
+    /// - the provided SMT does not track only valid [`Asset`]s.
+    /// - the vault key at which the asset is stored does not match the vault key derived from the
+    ///   asset.
+    pub fn new(partial_smt: PartialSmt) -> Result<Self, PartialAssetVaultError> {
+        Self::validate_entries(partial_smt.entries())?;
+
+        Ok(PartialVault { partial_smt })
     }
 
     // ACCESSORS
@@ -72,10 +83,47 @@ impl PartialVault {
     // MUTATORS
     // --------------------------------------------------------------------------------------------
 
-    // TODO: Do we need this method? It does not validate that the added value is a valid Asset.
     /// Adds an [`SmtProof`] to this [`PartialVault`].
-    pub fn add(&mut self, proof: SmtProof) -> Result<(), MerkleError> {
-        self.partial_smt.add_proof(proof)
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if:
+    /// - the provided proof does not prove inclusion of valid [`Asset`]s.
+    /// - the vault key of the proven asset does not match the vault key derived from the asset.
+    /// - the new root after the insertion of the leaf and the path does not match the existing root
+    ///   (except when the first leaf is added).
+    pub fn add(&mut self, proof: SmtProof) -> Result<(), PartialAssetVaultError> {
+        Self::validate_entries(proof.leaf().entries())?;
+
+        self.partial_smt
+            .add_proof(proof)
+            .map_err(PartialAssetVaultError::FailedToAddProof)
+    }
+
+    // HELPER FUNCTIONS
+    // --------------------------------------------------------------------------------------------
+
+    /// Validates that the provided entries are valid vault keys and assets.
+    ///
+    /// For brevity, the error conditions are only mentioned on the public methods that use this
+    /// function.
+    fn validate_entries<'a>(
+        entries: impl IntoIterator<Item = &'a (Word, Word)>,
+    ) -> Result<(), PartialAssetVaultError> {
+        for (vault_key, asset) in entries {
+            let asset = Asset::try_from(asset).map_err(|source| {
+                PartialAssetVaultError::InvalidAssetInSmt { entry: *asset, source }
+            })?;
+
+            if asset.vault_key() != *vault_key {
+                return Err(PartialAssetVaultError::VaultKeyMismatch {
+                    expected: asset.vault_key(),
+                    actual: *vault_key,
+                });
+            }
+        }
+
+        Ok(())
     }
 }
 
@@ -88,17 +136,71 @@ impl From<&AssetVault> for PartialVault {
 }
 
 impl Serializable for PartialVault {
-    fn write_into<W: vm_core::utils::ByteWriter>(&self, target: &mut W) {
+    fn write_into<W: ByteWriter>(&self, target: &mut W) {
         target.write(&self.partial_smt)
     }
 }
 
 impl Deserializable for PartialVault {
-    fn read_from<R: vm_core::utils::ByteReader>(
-        source: &mut R,
-    ) -> Result<Self, vm_processor::DeserializationError> {
+    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let vault_partial_smt = source.read()?;
 
-        Ok(PartialVault { partial_smt: vault_partial_smt })
+        PartialVault::new(vault_partial_smt)
+            .map_err(|err| DeserializationError::InvalidValue(err.to_string()))
+    }
+}
+
+// TESTS
+// ================================================================================================
+
+#[cfg(test)]
+mod tests {
+    use assert_matches::assert_matches;
+    use miden_crypto::merkle::Smt;
+
+    use super::*;
+    use crate::asset::FungibleAsset;
+
+    #[test]
+    fn partial_vault_ensures_asset_validity() -> anyhow::Result<()> {
+        let invalid_asset = Word::from([0, 0, 0, 5u32]);
+        let smt = Smt::with_entries([(invalid_asset, invalid_asset)])?;
+        let proof = smt.open(&invalid_asset);
+        let partial_smt = PartialSmt::from_proofs([proof.clone()])?;
+
+        let err = PartialVault::new(partial_smt).unwrap_err();
+        assert_matches!(err, PartialAssetVaultError::InvalidAssetInSmt { entry, .. } => {
+            assert_eq!(entry, invalid_asset);
+        });
+
+        let err = PartialVault::default().add(proof).unwrap_err();
+        assert_matches!(err, PartialAssetVaultError::InvalidAssetInSmt { entry, .. } => {
+            assert_eq!(entry, invalid_asset);
+        });
+
+        Ok(())
+    }
+
+    #[test]
+    fn partial_vault_ensures_asset_vault_key_matches() -> anyhow::Result<()> {
+        let asset = FungibleAsset::mock(500);
+        let invalid_vault_key = Word::from([0, 1, 2, 3u32]);
+        let smt = Smt::with_entries([(invalid_vault_key, asset.into())])?;
+        let proof = smt.open(&invalid_vault_key);
+        let partial_smt = PartialSmt::from_proofs([proof.clone()])?;
+
+        let err = PartialVault::new(partial_smt).unwrap_err();
+        assert_matches!(err, PartialAssetVaultError::VaultKeyMismatch { expected, actual } => {
+            assert_eq!(actual, invalid_vault_key);
+            assert_eq!(expected, asset.vault_key());
+        });
+
+        let err = PartialVault::default().add(proof).unwrap_err();
+        assert_matches!(err, PartialAssetVaultError::VaultKeyMismatch { expected, actual } => {
+            assert_eq!(actual, invalid_vault_key);
+            assert_eq!(expected, asset.vault_key());
+        });
+
+        Ok(())
     }
 }